DownloadDocmapIndex split code from FindNewDocmaps
