+++ starter/starter_FindNewDocmaps.py
@@ -1,4 +1,5 @@
 import json
+import uuid
 from starter.objects import Starter, default_workflow_params
 from provider import utils
 
@@ -13,23 +14,25 @@ class starter_FindNewDocmaps(Starter):
             settings, logger, "FindNewDocmaps"
         )
 
-    def get_workflow_params(self):
+    def get_workflow_params(self, run=None):
         workflow_params = default_workflow_params(self.settings)
         workflow_params["workflow_id"] = self.name
         workflow_params["workflow_name"] = self.name
         workflow_params["workflow_version"] = "1"
-        info = {}
+        info = {"run": run}
         workflow_params["input"] = json.dumps(info, default=lambda ob: None)
         return workflow_params
 
-    def start(self, settings):
+    def start(self, settings, run=None):
         "method for backwards compatibility"
         self.settings = settings
         self.instantiate_logger()
         self.start_workflow()
 
-    def start_workflow(self):
-        workflow_params = self.get_workflow_params()
+    def start_workflow(self, run=None):
+        if run is None:
+            run = str(uuid.uuid4())
+        workflow_params = self.get_workflow_params(run)
         self.start_workflow_execution(workflow_params)
